moved decoding for u's
