+++ xrpl-common-stdlib/src/fields/decoder.rs
@@ -32,6 +32,78 @@ pub trait FromCurrentTx: FieldDecoder {}
 /// Marker: this type can be read from a ledger object via [`crate::fields::ledger_obj`].
 pub trait FromLedger: FieldDecoder {}
 
+impl FieldDecoder for u8 {
+    type Buffer = [u8; 1];
+
+    #[inline]
+    fn empty_buffer() -> Self::Buffer {
+        [0u8; 1]
+    }
+
+    #[inline]
+    fn decode(bytes: &[u8]) -> Result<Self, DecodeError> {
+        let array: Self::Buffer = bytes.try_into().map_err(|_| DecodeError)?;
+        Ok(u8::from_le_bytes(array))
+    }
+}
+
+impl FromCurrentTx for u8 {}
+impl FromLedger for u8 {}
+
+impl FieldDecoder for u16 {
+    type Buffer = [u8; 2];
+
+    #[inline]
+    fn empty_buffer() -> Self::Buffer {
+        [0u8; 2]
+    }
+
+    #[inline]
+    fn decode(bytes: &[u8]) -> Result<Self, DecodeError> {
+        let array: Self::Buffer = bytes.try_into().map_err(|_| DecodeError)?;
+        Ok(u16::from_le_bytes(array))
+    }
+}
+
+impl FromCurrentTx for u16 {}
+impl FromLedger for u16 {}
+
+impl FieldDecoder for u32 {
+    type Buffer = [u8; 4];
+
+    #[inline]
+    fn empty_buffer() -> Self::Buffer {
+        [0u8; 4]
+    }
+
+    #[inline]
+    fn decode(bytes: &[u8]) -> Result<Self, DecodeError> {
+        let array: Self::Buffer = bytes.try_into().map_err(|_| DecodeError)?;
+        Ok(u32::from_le_bytes(array))
+    }
+}
+
+impl FromCurrentTx for u32 {}
+impl FromLedger for u32 {}
+
+impl FieldDecoder for u64 {
+    type Buffer = [u8; 8];
+
+    #[inline]
+    fn empty_buffer() -> Self::Buffer {
+        [0u8; 8]
+    }
+
+    #[inline]
+    fn decode(bytes: &[u8]) -> Result<Self, DecodeError> {
+        let array: Self::Buffer = bytes.try_into().map_err(|_| DecodeError)?;
+        Ok(u64::from_le_bytes(array))
+    }
+}
+
+impl FromCurrentTx for u64 {}
+impl FromLedger for u64 {}
+
 #[cfg(test)]
 mod tests {
     use super::*;
